chel-chatbot/app/requirements.txt
fastapi
uvicorn
python-dotenv
openai>=1.0.0
psycopg2-binary
